refs CRASM-2991: move sorting logic to reusable utils
- Created utils/sort.ts with naturalCompare and ipCompare functions
- Updated Domains.tsx to import and use sorting utilities for DataGrid columns
- Improves code maintainability and enables reuse in tests and other components
